Add files via upload
This version has the functions for creating the data storage array and the word search array. It has a method to fill the data array from user input. This method has a bug where it begins to fill the word search array instead of stopping after the data array has been filled. It has something to do with the inheritance or the for loop control. It currently does not perform as expected. Finally there is a for each loop method to print the data storage array to console.