Add files via upload
This update has a method to take a string from the data array and translate it into chars. Then it is placed in the 2d array "matrix" that will become the word search. There is currently an issue in this method as it throws a "Bad_Allocation" error. There is also a method to print out the word search. using nested for loops.