Add files via upload
This upload has a new method titled newArrange(). 
This method prints one word in the first row with a spacer too keep it from being left aligned. 
Then it calls compare() which takes a boolean and the i index of the first word (word[i]). Compare then reads both words and finds if the second word's first letter matches any letters in the first word. If so the indexes of each letter in their respective word is put into global array location[] and the boolean is set to true and returned to newArrange(). If not nothing happens
Back in newArrange() a second while loop will traverse the 2d array until it has found the column and row location of the common letter. Then the second word is spelled out downwards in that column by iterating the index value within that word (via location[]) and adding one to the row value. then it is printed as normal. Currently this only prints 2 words of the 4 in the test data. The next version should be able to print all four words although likely not in an overlapping manner.