Add files via upload
This update has a two new methods that arrange words horizontally and vertically. they are called arrangeHorizontal(i) and arrangeVertical(i) respectively. int i is the location of the word they arrange in the array list containing the words. They first search a row (or column) for empty spaces, if found that row (or column) number is saved and the word is printed in said space.  They are called from main and i is set manually. Currently the word search holds 6 words and is 12x12.